New button suggestions implemented
