+++ src/components/home/approach.tsx
@@ -6,7 +6,7 @@ import { GridContainer } from "../global/grid/gridContainer";
 import Pill from "../global/pill";
 import { motion } from "framer-motion";
 import { IconButton } from "../global/iconButton";
-import ArrowUpRight from "../../icons/arrowUpRight";
+import { PillIconButton } from "../global/pillIconButton";
 import Link from "next/link";
 import { getRemSize } from "../../styles/globalCss";
 import useIsIntersecting from "../../hooks/useIsIntersecting";
@@ -212,56 +212,56 @@ const SmallerIconButton = styled.div`
   }
 `;
 
-const StyledIcon = styled(ArrowUpRight)`
-  width: 30px;
-  height: 30px;
-  margin-left: -2px;
-  position: absolute;
-  top: 25%;
-  right: 6%;
+// const StyledIcon = styled(ArrowUpRight)`
+//   width: 30px;
+//   height: 30px;
+//   margin-left: -2px;
+//   position: absolute;
+//   top: 25%;
+//   right: 6%;
 
-  transition: 0.3s ease;
+//   transition: 0.3s ease;
 
+//   @media all and (max-width: ${breakpoints.md}px) {
+//     top: 38%;
+//     right: 6%;
+//   }
 
-  @media all and (max-width: ${breakpoints.md}px) {
-    top: 38%;
-    right: 6%;
-  }
+//   @media all and (max-width: ${breakpoints.sm}px) {
+//     top: 30%;
+//     right: 4%;
+// `;
 
-  @media all and (max-width: ${breakpoints.sm}px) {
-    top: 30%;
-    right: 4%;
-`;
+// const StyledPillButton = styled(motion.div)`
+//   width: 47%;
+//   max-width: 777px;
+//   min-width: 237px;
+//   margin-left: auto;
+//   margin-right: auto;
+//   margin-top: 24px;
+//   background-color: ${colors.white};
+//   color: ${colors.black};
+//   font-weight: 500;
+//   border-radius: 26px;
+//   padding: 4vw;
+//   font-size: 25px;
+//   letter-spacing: 2px;
+//   display: flex;
+//   align-items: center;
+//   position: relative;
+//   border: 2px solid rgba(29, 29, 29, 0.1);
 
-const StyledPillButton = styled(motion.div)`
-  width: 47%;
-  max-width: 777px;
-  min-width: 237px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 24px;
-  background-color: ${colors.white};
-  color: ${colors.black};
-  font-weight: 500;
-  border-radius: 26px;
-  padding: 4vw;
-  font-size: 25px;
-  letter-spacing: 2px;
-  display: flex;
-  align-items: center;
-  position: relative;
-  border: 2px solid rgba(29, 29, 29, 0.1);
+//   &:hover .styled-icon {
+//     transform: rotate(45deg);
+//   }
 
-  &:hover .styled-icon {
-    transform: rotate(45deg);
-  }
+//   @media all and (max-width: ${breakpoints.sm}px) {
+//     min-width: 237px;
+//     padding: 4vw;
+//     font-size: 25px;
+//   }
+// `;
 
-  @media all and (max-width: ${breakpoints.sm}px) {
-    min-width: 237px;
-    padding: 4vw;
-    font-size: 25px;
-  }
-`;
 interface StyledCardProps {
   marginLeft?: string;
 }
@@ -326,12 +326,7 @@ export default function Approach({ items }: IApproach) {
               </SmallerIconButton>
             </StyledCardPill>
           ) : (
-            <StyledPillButton>
-              Check out more
-              <Link href="/projects">
-                <StyledIcon className="styled-icon" />
-              </Link>
-            </StyledPillButton>
+            <PillIconButton text="Check out more" />
           )}
         </StyledBox>
         <StyledApproachBorderRight
